requirements.txt
# LangChain / LangGraph
langchain
langgraph
langchain-anthropic
langchain-chroma
langgraph-cli
langdetect

# Vector store
chromadb

# Embeddings
openai
tiktoken

# PDF parsing
pdfplumber

# Scraping
beautifulsoup4
requests

# Data handling
pydantic

# Environment
python-dotenv

# Evaluation
pandas